map_my_crop/requirements.txt
asgiref==3.7.2
attrs==23.2.0
cattrs==23.2.3
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
flatbuffers==23.5.26
idna==3.6
numpy==1.26.3
openmeteo_requests==1.1.0
openmeteo_sdk==1.7.2
pandas==2.1.4
platformdirs==4.1.0
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-cache==1.1.1
retry-requests==2.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
url-normalize==1.4.3
urllib3==2.1.0
